5386: working on vlaidation [skip ci]
