+++ web-client/src/getConstants.js
@@ -18,6 +18,7 @@ import {
   CHAMBERS_SECTIONS_LABELS,
   CHIEF_JUDGE,
   CONTACT_CHANGE_DOCUMENT_TYPES,
+  CONTACT_TYPES,
   COUNTRY_TYPES,
   COURT_ISSUED_DOCUMENT_TYPES,
   COURT_ISSUED_EVENT_CODES,
@@ -119,6 +120,7 @@ export const getConstants = () => ({
   CHANNEL_NAME: 'ustc-broadcast',
   CHIEF_JUDGE,
   CONTACT_CHANGE_DOCUMENT_TYPES,
+  CONTACT_TYPES,
   COUNTRY_TYPES,
   COURT_ISSUED_DOCUMENT_TYPES,
   COURT_ISSUED_EVENT_CODES,
