+++ web-client/src/presenter/actions/setDefaultAddPetitionerToCaseFormAction.js
@@ -1,3 +1,4 @@
+import { getContactSecondary } from '../../../../shared/src/business/entities/cases/Case';
 import { state } from 'cerebral';
 
 /**
@@ -6,6 +7,21 @@ import { state } from 'cerebral';
  * @param {object} providers the providers object
  * @param {object} providers.store the cerebral store
  */
-export const setDefaultAddPetitionerToCaseFormAction = ({ store }) => {
+export const setDefaultAddPetitionerToCaseFormAction = ({
+  applicationContext,
+  get,
+  store,
+}) => {
+  const { CONTACT_TYPES } = applicationContext.getConstants();
+
   store.set(state.form, { contact: {} });
+
+  const caseDetail = get(state.caseDetail);
+  const contactSecondary = getContactSecondary(caseDetail);
+
+  if (!contactSecondary) {
+    store.set(state.form.contactType, CONTACT_TYPES.secondary);
+  } else {
+    store.set(state.form.contactType, CONTACT_TYPES.otherPetitioner);
+  }
 };
